+++ history.md
@@ -42,4 +42,5 @@
 - 1.0.0-preview5 (unreleased)
   - *added:* `Transient` services are freed on injector disposal (if they have not been disposed programmatically)
   - *added:* `ObjectExtensions.Act()` can be called against interfaces
+  - *added:* `ProxyGenerator` now supports `in` parameter modifier (introduced in C# 7.2)
   - *renamed:* `QueryOptionsAttribute` to `OptionsAttribute`